+++ auditlog/apps.py
@@ -1,9 +1,10 @@
 from django.apps import AppConfig
+from django.utils.translation import gettext_lazy as _
 
 
 class AuditlogConfig(AppConfig):
     name = "auditlog"
-    verbose_name = "Audit log"
+    verbose_name = _("Audit log")
     default_auto_field = "django.db.models.AutoField"
 
     def ready(self):