+++ auditlog/middleware.py
@@ -1,5 +1,6 @@
 import contextlib
 
+from auditlog.cid import set_cid
 from auditlog.context import set_actor
 
 
@@ -12,13 +13,27 @@ class AuditlogMiddleware:
     def __init__(self, get_response=None):
         self.get_response = get_response
 
-    def __call__(self, request):
+    @staticmethod
+    def _get_remote_addr(request):
+        # In case there is no proxy, return the original address
+        if not request.headers.get("X-Forwarded-For"):
+            return request.META.get("REMOTE_ADDR")
 
-        if request.META.get("HTTP_X_FORWARDED_FOR"):
-            # In case of proxy, set 'original' address
-            remote_addr = request.META.get("HTTP_X_FORWARDED_FOR").split(",")[0]
-        else:
-            remote_addr = request.META.get("REMOTE_ADDR")
+        # In case of proxy, set 'original' address
+        remote_addr: str = request.headers.get("X-Forwarded-For").split(",")[0]
+
+        # Remove port number from remote_addr
+        if "." in remote_addr and ":" in remote_addr:  # IPv4 with port (`x.x.x.x:x`)
+            remote_addr = remote_addr.split(":")[0]
+        elif "[" in remote_addr:  # IPv6 with port (`[:::]:x`)
+            remote_addr = remote_addr[1:].split("]")[0]
+
+        return remote_addr
+
+    def __call__(self, request):
+        remote_addr = self._get_remote_addr(request)
+
+        set_cid(request)
 
         if hasattr(request, "user") and request.user.is_authenticated:
             context = set_actor(actor=request.user, remote_addr=remote_addr)