+++ auditlog/mixins.py
@@ -1,26 +1,34 @@
-import json
-
 from django import urls as urlresolvers
 from django.conf import settings
+from django.contrib import admin
+from django.core.exceptions import FieldDoesNotExist
 from django.db.models import DateTimeField
+from django.forms.utils import pretty_name
+from django.http import HttpRequest
 from django.template.defaultfilters import pluralize
 from django.urls.exceptions import NoReverseMatch
 from django.utils import timezone
 from django.utils.html import format_html, format_html_join
 from django.utils.safestring import mark_safe
 from django.utils.timezone import localtime
+from django.utils.translation import gettext_lazy as _
 
 from auditlog.models import LogEntry
+from auditlog.registry import auditlog
+from auditlog.signals import accessed
 
 MAX = 75
 
 
 class LogEntryAdminMixin:
+    request: HttpRequest
+    CID_TITLE = _("Click to filter by records with this correlation id")
+
+    @admin.display(description=_("Created"))
     def created(self, obj):
         return localtime(obj.timestamp).strftime("%Y-%m-%d %H:%M:%S")
 
-    created.short_description = "Created"
-
+    @admin.display(description=_("User"))
     def user_url(self, obj):
         if obj.actor:
             app_label, model = settings.AUTH_USER_MODEL.split(".")
@@ -33,8 +41,7 @@ class LogEntryAdminMixin:
 
         return "system"
 
-    user_url.short_description = "User"
-
+    @admin.display(description=_("Resource"))
     def resource_url(self, obj):
         app_label, model = obj.content_type.app_label, obj.content_type.model
         viewname = f"admin:{app_label}_{model}_change"
@@ -48,12 +55,11 @@ class LogEntryAdminMixin:
                 '<a href="{}">{} - {}</a>', link, obj.content_type, obj.object_repr
             )
 
-    resource_url.short_description = "Resource"
-
+    @admin.display(description=_("Changes"))
     def msg_short(self, obj):
-        if obj.action == LogEntry.Action.DELETE:
+        if obj.action in [LogEntry.Action.DELETE, LogEntry.Action.ACCESS]:
             return ""  # delete
-        changes = json.loads(obj.changes)
+        changes = obj.changes_dict
         s = "" if len(changes) == 1 else "s"
         fields = ", ".join(changes.keys())
         if len(fields) > MAX:
@@ -61,12 +67,9 @@ class LogEntryAdminMixin:
             fields = fields[:i] + " .."
         return "%d change%s: %s" % (len(changes), s, fields)
 
-    msg_short.short_description = "Changes"
-
+    @admin.display(description=_("Changes"))
     def msg(self, obj):
-        if obj.action == LogEntry.Action.DELETE:
-            return ""  # delete
-        changes = json.loads(obj.changes)
+        changes = obj.changes_dict
 
         atom_changes = {}
         m2m_changes = {}
@@ -88,7 +91,9 @@ class LogEntryAdminMixin:
             msg.append("<table>")
             msg.append(self._format_header("#", "Field", "From", "To"))
             for i, (field, change) in enumerate(sorted(atom_changes.items()), 1):
-                value = [i, field] + (["***", "***"] if field == "password" else change)
+                value = [i, self.field_verbose_name(obj, field)] + (
+                    ["***", "***"] if field == "password" else change
+                )
                 if field in datetime_fields:
                     spotted_datetime_field = True
                 msg.append(self._format_line(*value))
@@ -108,7 +113,7 @@ class LogEntryAdminMixin:
                     format_html(
                         "<tr><td>{}</td><td>{}</td><td>{}</td><td>{}</td></tr>",
                         i,
-                        field,
+                        self.field_verbose_name(obj, field),
                         change["operation"],
                         change_html,
                     )
@@ -123,8 +128,6 @@ class LogEntryAdminMixin:
 
         return mark_safe("".join(msg))
 
-    msg.short_description = "Changes"
-
     def _get_datetime_fields(self, obj):
         # only works for existing models and existing fields
         try:
@@ -135,6 +138,15 @@ class LogEntryAdminMixin:
             }
         except Exception:
             return set()
+
+    @admin.display(description="Correlation ID")
+    def cid_url(self, obj):
+        cid = obj.cid
+        if cid:
+            url = self._add_query_parameter("cid", cid)
+            return format_html(
+                '<a href="{}" title="{}">{}</a>', url, self.CID_TITLE, cid
+            )
 
     def _format_header(self, *labels):
         return format_html(
@@ -162,3 +174,32 @@ class LogEntryAdminMixin:
             )
         )
         return '<span class="timezonewarning">{}</span>'.format(warning_message)
+
+    def field_verbose_name(self, obj, field_name: str):
+        model = obj.content_type.model_class()
+        try:
+            model_fields = auditlog.get_model_fields(model._meta.model)
+            mapping_field_name = model_fields["mapping_fields"].get(field_name)
+            if mapping_field_name:
+                return mapping_field_name
+        except KeyError:
+            # Model definition in auditlog was probably removed
+            pass
+        try:
+            field = model._meta.get_field(field_name)
+            return pretty_name(getattr(field, "verbose_name", field_name))
+        except FieldDoesNotExist:
+            return pretty_name(field_name)
+
+    def _add_query_parameter(self, key: str, value: str):
+        full_path = self.request.get_full_path()
+        delimiter = "&" if "?" in full_path else "?"
+
+        return f"{full_path}{delimiter}{key}={value}"
+
+
+class LogAccessMixin:
+    def render_to_response(self, context, **response_kwargs):
+        obj = self.get_object()
+        accessed.send(obj.__class__, instance=obj)
+        return super().render_to_response(context, **response_kwargs)