+++ auditlog/receivers.py
@@ -1,9 +1,31 @@
 import json
+from functools import wraps
 
+from django.conf import settings
+
+from auditlog.context import threadlocal
 from auditlog.diff import model_instance_diff
 from auditlog.models import LogEntry
 
 
+def check_disable(signal_handler):
+    """
+    Decorator that passes along disabled in kwargs if any of the following is true:
+    - 'auditlog_disabled' from threadlocal is true
+    - raw = True and AUDITLOG_DISABLE_ON_RAW_SAVE is True
+    """
+
+    @wraps(signal_handler)
+    def wrapper(*args, **kwargs):
+        if not getattr(threadlocal, "auditlog_disabled", False) and not (
+            kwargs.get("raw") and settings.AUDITLOG_DISABLE_ON_RAW_SAVE
+        ):
+            signal_handler(*args, **kwargs)
+
+    return wrapper
+
+
+@check_disable
 def log_create(sender, instance, created, **kwargs):
     """
     Signal receiver that creates a log entry when a model instance is first saved to the database.
@@ -20,6 +42,7 @@ def log_create(sender, instance, created, **kwargs):
         )
 
 
+@check_disable
 def log_update(sender, instance, **kwargs):
     """
     Signal receiver that creates a log entry when a model instance is changed and saved to the database.
@@ -45,6 +68,7 @@ def log_update(sender, instance, **kwargs):
                 )
 
 
+@check_disable
 def log_delete(sender, instance, **kwargs):
     """
     Signal receiver that creates a log entry when a model instance is deleted from the database.
@@ -61,9 +85,25 @@ def log_delete(sender, instance, **kwargs):
         )
 
 
+def log_access(sender, instance, **kwargs):
+    """
+    Signal receiver that creates a log entry when a model instance is accessed in a AccessLogDetailView.
+
+    Direct use is discouraged, connect your model through :py:func:`auditlog.registry.register` instead.
+    """
+    if instance.pk is not None:
+
+        LogEntry.objects.log_create(
+            instance,
+            action=LogEntry.Action.ACCESS,
+            changes="null",
+        )
+
+
 def make_log_m2m_changes(field_name):
     """Return a handler for m2m_changed with field_name enclosed."""
 
+    @check_disable
     def log_m2m_changes(signal, action, **kwargs):
         """Handle m2m_changed and call LogEntry.objects.log_m2m_changes as needed."""
         if action not in ["post_add", "post_clear", "post_remove"]: